+++ Kernel/System/DynamicField/Driver/Date.pm
@@ -396,6 +396,25 @@ sub DisplayValueRender {
     return $Data;
 }
 
+sub ReadableValueRender {
+    my ( $Self, %Param ) = @_;
+
+    my $Value = defined $Param{Value} ? $Param{Value} : '';
+
+    # only keep date part, loose time part of timestamp
+    $Value =~ s/^(\d\d\d\d-\d\d-\d\d).+?$/$1/xms;
+
+    # Title is always equal to Value
+    my $Title = $Value;
+
+    my $Data = {
+        Value => $Value,
+        Title => $Title,
+    };
+
+    return $Data;
+}
+
 sub SearchFieldRender {
     my ( $Self, %Param ) = @_;