+++ Kernel/System/DynamicField/Driver/Multiselect.pm
@@ -530,9 +530,13 @@ sub SearchFieldParameterBuild {
 
                 # set the display value
                 my $DisplayItem = $Param{DynamicFieldConfig}->{Config}->{PossibleValues}->{$Item};
-                if ( $Param{DynamicFieldConfig}->{Config}->{TranslatableValues} ) {
 
-                    # translate the value
+                # translate the value
+                if (
+                    $Param{DynamicFieldConfig}->{Config}->{TranslatableValues}
+                    && defined $Param{LayoutObject}
+                    )
+                {
                     $DisplayItem = $Param{LayoutObject}->{LanguageObject}->Get($DisplayItem);
                 }
 
@@ -547,9 +551,12 @@ sub SearchFieldParameterBuild {
             # set the display value
             $DisplayValue = $Param{DynamicFieldConfig}->{PossibleValues}->{$Value};
 
-            if ( $Param{DynamicFieldConfig}->{Config}->{TranslatableValues} ) {
-
-                # translate the value
+            # translate the value
+            if (
+                $Param{DynamicFieldConfig}->{Config}->{TranslatableValues}
+                && defined $Param{LayoutObject}
+                )
+            {
                 $DisplayValue = $Param{LayoutObject}->{LanguageObject}->Get($DisplayValue);
             }
         }
@@ -793,7 +800,7 @@ sub BuildSelectionDataGet {
                 if ( !$Key && !$PosibleNoneSet && $FieldConfig->{PossibleNone} ) {
 
                     my $Selected;
-                    if ( 
+                    if (
                         !IsHashRefWithData(\%Values)
                         || ( defined $Values{''} &&  $Values{''} )
                         )